refactor(utils): create gimie.utils.uri submodule
